+++ tests/tests_unit/test_api/test_functions.py
@@ -185,6 +185,16 @@ def function_handle_illegal_argument():
     return handle
 
 
+@pytest.fixture
+def mock_function_calls_list_response(rsps):
+
+    response_body = {"items": [CALL_COMPLETED]}
+    url = FUNCTIONS_API._get_base_url_with_base_path() + f"/functions/{FUNCTION_ID}/calls/list"
+    rsps.add(rsps.POST, url, status=200, json=response_body)
+
+    yield rsps
+
+
 class TestFunctionsAPI:
     @pytest.mark.parametrize(
         "function_folder, will_pass",
@@ -302,15 +312,6 @@ class TestFunctionsAPI:
         res = FUNCTIONS_API.call(id=FUNCTION_ID)
         assert isinstance(res, FunctionCall)
         assert mock_functions_call_timeout_response.calls[0].response.json() == res.dump(camel_case=True)
-
-
-@pytest.fixture
-def mock_function_calls_list_response(rsps):
-    response_body = {"items": [CALL_COMPLETED]}
-    url = FUNCTIONS_API._get_base_url_with_base_path() + f"/functions/{FUNCTION_ID}/calls"
-    rsps.add(rsps.GET, url, status=200, json=response_body)
-
-    yield rsps
 
 
 @pytest.fixture
@@ -435,8 +436,26 @@ class TestFunctionSchedulesAPI:
 
 
 class TestFunctionCallsAPI:
+    def test_list_calls_and_filter(self, mock_function_calls_list_response, mock_functions_retrieve_response):
+        filter_kwargs = {
+            "status": "Completed",
+            "schedule_id": 123,
+            "start_time": {"min": 1585925306822, "max": 1585925306823},
+            "end_time": {"min": 1585925310822, "max": 1585925310823},
+        }
+        res = FUNCTIONS_API.retrieve(id=FUNCTION_ID).list_calls(**filter_kwargs)
+
+        assert isinstance(res, FunctionCallList)
+        assert mock_function_calls_list_response.calls[1].response.json()["items"] == res.dump(camel_case=True)
+
     def test_list_calls_by_function_id(self, mock_function_calls_list_response):
-        res = FUNCTION_CALLS_API.list(function_id=FUNCTION_ID)
+        filter_kwargs = {
+            "status": "Completed",
+            "schedule_id": 123,
+            "start_time": {"min": 1585925306822, "max": 1585925306823},
+            "end_time": {"min": 1585925310822, "max": 1585925310823},
+        }
+        res = FUNCTION_CALLS_API.list(function_id=FUNCTION_ID, **filter_kwargs)
         assert isinstance(res, FunctionCallList)
         assert mock_function_calls_list_response.calls[0].response.json()["items"] == res.dump(camel_case=True)